init 3

+++ MyLibrary/AverageCalculator.cs
@@ -6,7 +6,7 @@ namespace MyLibrary
 {
     public class AverageCalculator
     {
-        public double calculate(int totalMarksValue, int totalMarksImpact)
+        public double calculate(double totalMarksValue, int totalMarksImpact)
         {
             if(totalMarksImpact != 0 && totalMarksValue != 0)
             {
